+++ jest.config.js
@@ -158,6 +158,7 @@ module.exports = {
 
   // The regexp pattern or array of patterns that Jest uses to detect test files
   testRegex: [
+    // "base\.test\.js",
     // "other\.test\.js",
   ],